Organize the python folder structure
The folder structure is now more similar to a package structure, since
this allows the project to be installed locally and make the import in
tests and reports a more doable process than making a relative path to
import or the commands to run the tests be more concise and less
confused.

As it goes for the structure itself, the structure has its base on the
`/rman/` directory, followed by three sub-packages: `/dhc/`, `/kinematic/` and
`/dynamic/`. In these, the different modules will be developed to enable
the control of the entire robotic arm. For now, only the dhc and kinematic
sub-packages are in development, with direct already concluded and
differential in the testing process.
